ai_solutions_agency_os-2/requirements.txt
langgraph>=0.2.0
pydantic>=2.7.0
python-dotenv>=1.0.1
openai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
